Add the concept of a "Runbook".
This is an object that can be added to a component, or an individual report. It should contain a URL describing where a person responding to the alert could go to find information on how to respond to an alert: who to contact; recovery time objective; investigation steps; etc.

If added to a component, it will be used when an exception is thrown outside of getReport.
If added to a component, it will be overridden by the report's Runbook, if present.