Support kubernetes readiness checks
In Kubernetes a readiness check uses the status code of the response instead
of the content as Tucker had been designed. A status 400 > CODE >= 200 indicates
that the container is ready.

This changes Tucker to support a "ready" resource that responds 204 for "healthy"
and 503 (Service Unavailable) for "ill", which makes it able to be used as
a Kubernetes readiness check.