Include Status Page compoents as metrics
In order to allow alerting based on status pages when using prometheus
and alertmanager the status page components need to be in the metrics
somehow. The scheme used here is that each (component, status) combination
is a label on the tucker_component_status metric. The pair that has the
current status of the component is set to 1. All other status pairs for
the component are set to 0.