Close output stream after handling metrics request
The metrics handler did not close the output stream after having written
the metrics. This caused, for the test server, the connection to hang open
until the other end timed out.

I believe the this problem didn't affect the Jetty servers using this, but
this now makes it consistent handling across all of the different handlers.

The flush is no longer needed as the data will get written to the client
by the close.