mpimm: Prepare tucker for having a metrics writer plugged in
How metrics are written will now be the concern of tim-metrics.

To maintain isolation between this tim library and tucker this adds the
MetricWriter interface which abstracts the writing of metrics.

In order to make way for this some Writers renamed to handlers

In order to make the constructor chaos a bit more manageable builder
added.